requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow==10.4.0
PyPDF2==3.0.1
pdf2docx==0.5.6
python-docx==1.1.2
PyMuPDF==1.24.9
camelot-py==0.11.0
pytesseract==0.3.13
openpyxl==3.1.5
gunicorn==23.0.0